refactor(engine): extract the git-evidence substrate from self-healing.ts (U4 decomposition, PR1)
PURE RELOCATION. 19 blocks moved to `self-healing-git-evidence.ts`: the 14
git-evidence readers, four module-level helpers they need, and the
`LandedTaskCommit` interface.

MECHANICAL PROOF, which is the point of this PR: every moved body was diffed
against its pre-move text. 19/19 are BYTE-IDENTICAL modulo the three deviations
enumerated below. No condition reordered, no signature changed, no rename, no
inlining, no "while I am here" cleanup.

THE PREMISE NEEDED CORRECTING, and this is the cheap lesson the first cut was
supposed to buy. The brief described 13 pure functions; they are 13 `private
async` METHODS closing over `this.options`, so no extraction is byte-identical in
the strict sense. Three deviations were structurally unavoidable:

  1. `private` -> `protected` on the 14 methods and the `options` field, because a
     subclass cannot call a `private` base member. One token per declaration.
  2. Two `SelfHealingManager["readCommitTaskOwnership"]` type annotations rewritten
     to name the base class; the original would be a circular import.
  3. Four module-level helpers moved with the methods and re-exported. This was
     forced by direction: the new module must not import self-healing.ts, so
     everything the bodies call has to live beside them. `execAsync` and
     `shellQuote` are imported back because call sites remain behind.

THE SET IS 14, NOT 13. `foreignTipRejection` had to come too, and it is what
CLOSES the cluster — it depends only on `baseHasExplicitTaskOwnership` and
`branchHasNoUniqueDiff`, both already in the set. Without it the extraction is not
self-contained and `this.options` would not be the only external dependency.

A BASE CLASS RATHER THAN FREE FUNCTIONS, deliberately. Converting these to free
functions would change all 14 signatures — a behavior-adjacent edit riding inside
a file move, which is exactly the combination that hid the last four safeguard
regressions. An abstract base keeps `this` semantics so every call site stays
`this.<method>(...)` and every body is unchanged text.

Two collisions worth knowing about, both PRE-EXISTING and preserved exactly: the
detector module already exports a FREE `findAlreadyMergedTaskCommit` that shares
its name with the protected method, and inside the class body the bare identifier
resolves to the import. And `SelfHealingOptions` is declared in self-healing.ts,
imported here TYPE-ONLY so it is erased at runtime and creates no module cycle.

MEASURED LINE DELTA, not an estimate:

  self-healing.ts   13,394 -> 12,945   = -449
  new module                           = +527   (454 moved verbatim, 73 scaffold)
  NET                                  = +78

Same shape as every other consolidation in this program: the target file shrinks,
the total grows slightly. At -449 for the first and safest cut, the remaining
substrate cuts plausibly take self-healing.ts under 11k, but the operator should
read that as file-size reduction rather than code reduction.

Verified: 464 passed across three engine suites with the single known
PRE-EXISTING `archiveStaleDoneTasks` failure, tsc clean, lint clean, merge gate
green (299+10+71).

Fusion-Task-Id: KB-U4
